// libs/core/src/model/interfaces/ICostOvertime.ts
// 9fbef606107a605d69c0edbcd8029e5d_SYMPHONY

export interface ICostData {
  date: string;
  averageCostPerShipment: string | number;
  totalcustomerCharge: string | number;
  averageCostPerUnit: string | number;
  averageCostPerMile: string | number;
  averageCostPerWeight: string | number;
  averageCostPerSKU: string | number;
  shipmentMode: any;
}